refactor: make AppHeader a server component
